+++ CHANGELOG.md
@@ -49,6 +49,24 @@
   + class MyRows extends GridAxis<T> with RowsMixin<T> { /* ... */ }
   ```
 
+- Changed `<Grid>.traversal` and `Traversal` to allow arbitrary return types,
+  i.e. not strictly a `GridIterable<T>`. This allows users to use the traversal
+  API to create custom traversals that return different types of elements, e.g.
+  custom transformations, filters, or other types of elements:
+
+  ```diff
+  - Traversal<T> rowMajor<T>(...) { ... }
+  + Traversal<GridIterable<T>, T> rowMajor<T>(...) { ... }
+  ```
+
+  As a result, `.traversal` now _requires_ an argument, versus defaulting to
+  `rowMajor()` before:
+
+  ```diff
+  - for (final element in grid.traversal) { /* ... */ }
+  + for (final element in grid.traversal(rowMajor())) { /* ... */ }
+  ```
+  
 ### New Features
 
 - Extending or mixing in `Grid<T>` provides a default implementation of `clear`.